requirements.txt
openai-agents>=0.0.11
python-dotenv>=1.0.0
pydantic>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dateutil>=2.8.2
pytz>=2024.1 